+++ mapie/conformity_scores/__init__.py
@@ -1,11 +1,10 @@
 from .bounds import (
     AbsoluteConformityScore,
     GammaConformityScore,
-    MultivariateResidualNormalisedScore,
     ResidualNormalisedScore,
 )
 from .classification import BaseClassificationScore
-from .regression import BaseFitRegressionScore, BaseRegressionScore
+from .regression import BaseRegressionScore
 from .sets import (
     APSConformityScore,
     LACConformityScore,
@@ -16,7 +15,6 @@ from .sets import (
 
 __all__ = [
     "BaseRegressionScore",
-    "BaseFitRegressionScore",
     "BaseClassificationScore",
     "AbsoluteConformityScore",
     "GammaConformityScore",
@@ -26,5 +24,4 @@ __all__ = [
     "APSConformityScore",
     "RAPSConformityScore",
     "TopKConformityScore",
-    "MultivariateResidualNormalisedScore",
 ]
