+++ mapie/conformity_scores/bounds/__init__.py
@@ -1,11 +1,9 @@
 from .absolute import AbsoluteConformityScore
 from .gamma import GammaConformityScore
 from .residuals import ResidualNormalisedScore
-from .standardized_residuals import MultivariateResidualNormalisedScore
 
 __all__ = [
     "AbsoluteConformityScore",
     "GammaConformityScore",
     "ResidualNormalisedScore",
-    "MultivariateResidualNormalisedScore",
 ]
