+++ mapie/experimental/__init__.py
@@ -1,0 +1,9 @@
+from mapie.experimental._base import BaseFitRegressionScore
+from mapie.experimental.standardized_residuals import (
+    MultivariateResidualNormalisedScore,
+)
+
+__all__ = [
+    "BaseFitRegressionScore",
+    "MultivariateResidualNormalisedScore",
+]
